1.5 second delay
A bit of a hack to add about 2s delay in the ISR.  The timing seems to be wrong based on the existing calculation -- need to verify the oscillator speed.  Still also need to figure out why the debugger reports a memory write failure due to the device description file.
Also updated .gitignore to drop the Debug folder.